数据处理完成

+++ model/teacher_schedule.go
@@ -1,6 +1,6 @@
 package model
 
 type TeacherSchedule struct {
-	TeacherId uint16
+	TeacherId int64
 	Schedule  []string // 这里存放class  class格式为 Day + starttime + endtime
 }
